Start of reorg. Root views w/ level being the determinator in App
